Backend-Systems/AsyncDB-FastAPI/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
motor==4.3.1
pydantic==2.6.2
python-jose==3.3.0
passlib[bcrypt]==1.7.5